SRAM Memory Writing
Have include RAM and Time modules, code compiles.

When run in debug the code will read the memory, write to the memory and then reread memory as a different value. Due to debugging resetting all SRAM it's not possible to see in debug if it reads a different value on power cycle.